requirements.txt
functions-framework==3.*
requests==2.31.0
pyarrow==12.0.0
fastparquet==2023.4.0
coverage==7.0.3
future==0.18.2
pytest==7.2.0
pytest-cov==4.0.0
flake8==6.0.0
pydocstyle==6.3.0
mock==5.0.2
pg8000==1.20.0
pytz==2023.3
psycopg2-binary==2.9.7
google-cloud-pubsub==2.18.3

